EvoNN-Compare/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evonn-compare"
version = "0.1.0"
description = "Protocol-first comparison and hybrid orchestration for Prism, Topograph, and their fusion"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.10.0",
    "pyyaml>=6.0.2",
    "scipy>=1.14.0",
    "typer>=0.15.1",
    "rich>=13.9.0",
    "numpy>=2.2.0",
    "scikit-learn>=1.6.0",
    "openml>=0.14.2",
    "pandas>=2.2.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
evonn-compare = "evonn_compare.cli.main:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
